Move updating dataset store out of trace mapping and make the set mappings non-specific to be used later on evaluations too
