+++ langwatch/src/components/datasets/DatasetMappingPreview.tsx
@@ -12,9 +12,11 @@ import {
 
 import type { CustomCellRendererProps } from "@ag-grid-community/react";
 import type { Dataset } from "@prisma/client";
-import { useMemo } from "react";
+import { useCallback, useMemo } from "react";
 import { Checkbox } from "../../components/ui/checkbox";
 import { TracesMapping, type MappingState } from "../traces/TracesMapping";
+import { api } from "../../utils/api";
+import { useOrganizationTeamProject } from "../../hooks/useOrganizationTeamProject";
 
 interface DatasetMappingPreviewProps {
   traces: any[]; // Replace 'any' with your trace type
@@ -81,6 +83,31 @@ export function DatasetMappingPreview({
     return headers;
   }, [selectedDataset]);
 
+  const { project } = useOrganizationTeamProject();
+
+  const trpc = api.useContext();
+  const updateStoredMapping_ = api.dataset.updateMapping.useMutation();
+  const updateStoredMapping = useCallback(
+    (mappingState: MappingState) => {
+      updateStoredMapping_.mutate(
+        {
+          projectId: project?.id ?? "",
+          datasetId: selectedDataset.id,
+          mapping: {
+            mapping: mappingState.mapping,
+            expansions: Array.from(mappingState.expansions),
+          },
+        },
+        {
+          onSuccess: () => {
+            void trpc.dataset.getAll.invalidate();
+          },
+        }
+      );
+    },
+    [selectedDataset.id, project?.id, trpc.dataset.getAll, updateStoredMapping_]
+  );
+
   return (
     <Field.Root width="full" paddingY={4}>
       <HStack width="full" gap="64px" align="start">
@@ -95,7 +122,10 @@ export function DatasetMappingPreview({
             traces={traces}
             columnTypes={columnTypes}
             setDatasetEntries={onRowDataChange}
-            setDatasetTriggerMapping={setDatasetTriggerMapping}
+            setDatasetMapping={(newMappingState) => {
+              setDatasetTriggerMapping?.(newMappingState);
+              updateStoredMapping(newMappingState);
+            }}
           />
         </VStack>
         <VStack align="start" width="full" height="full">
